+++ lib/assent/strategies/azure_ad.ex
@@ -7,8 +7,6 @@ defmodule Assent.Strategy.AzureAD do
   - `:client_id` - The OAuth2 client id, required
   - `:tenant_id` - The Azure tenant ID, optional, defaults to `common`
   - `:nonce` - The session based nonce, required
-  - `:resource` - The Azure resource, optional, defaults to
-    `https://graph.microsoft.com/`
 
   See `Assent.Strategy.OIDC` for more.
 
@@ -44,31 +42,13 @@ defmodule Assent.Strategy.AzureAD do
         tenant_id: "REPLACE_WITH_TENANT_ID"
       ]
 
-  The resource that client should pull a token for defaults to
-  `https://graph.microsoft.com/`. It can be overridden with the
-  `resource` key (or the `authorization_params` key):
-
-      config = [
-        client_id: "REPLACE_WITH_CLIENT_ID",
-        nonce: "DYNAMICALLY_REPLACE_WITH_SESSION_NONCE",
-        tenant_id: "REPLACE_WITH_TENANT_ID",
-        resource: "https://service.contoso.com/"
-      ]
-
   ## Setting up Azure AD
 
   Login to Azure, and set up a new application:
-  https://docs.microsoft.com/en-us/azure/active-directory/develop/v1-protocols-openid-connect-code#register-your-application-with-your-ad-tenant
+  https://docs.microsoft.com/en-us/azure/active-directory/develop/quickstart-register-app
 
+  * The callback URL should be added to "Redirect URI" for the application.
   * `client_id` is the "Application ID".
-  * The callback URL should be added to Redirect URI for the application.
-  * "Sign in and read user profile" permission has to be enabled.
-
-  ### App ID URI for `resource`
-
-  To find the App ID URI to be used for `resource`, in the Azure Portal, click
-  Azure Active Directory, click Application registrations, open the
-  application's Settings page, then click Properties.
   """
   use Assent.Strategy.OIDC.Base
 
@@ -77,12 +57,11 @@ defmodule Assent.Strategy.AzureAD do
   @impl true
   def default_config(config) do
     tenant_id = Config.get(config, :tenant_id, "common")
-    resource  = Config.get(config, :resource, "https://graph.microsoft.com/")
 
     [
       site: "https://login.microsoftonline.com/#{tenant_id}/v2.0",
-      authorization_params: [response_type: "id_token code", resource: resource],
-      client_auth_method: :client_secret_post
+      authorization_params: [response_type: "id_token code", scope: "email profile", response_mode: "form_post"],
+      client_auth_method: :client_secret_post,
     ]
   end