+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v0.1.5 (TBA)
+
+* Removed unused `:resource` param in `Assent.Strategy.AzureAD`
+* Added "email profile" to scope in `Assent.Strategy.AzureAD`
+* Use `response_mode=form_post` for `Assent.Strategy.AzureAD`
+
 ## v0.1.4 (2019-11-09)
 
 * Support mint up to `v1.0.x`